requirements.txt
requests==2.31.0
numpy==1.24.4
pillow==10.4.0